requirements.txt
# flask and flask plugins
Flask>=0.10
Flask-Admin>=1.0.7
Flask-Bcrypt
Flask-Bootstrap
Flask-Cache
Flask-Login==0.2.11
pymongo==2.6.2
mongoengine==0.8.4
Flask-MongoEngine==0.7.0
Flask-Reggie
Flask-SeaSurf
Flask-SSLify
Flask-Views
Flask-WTF>=0.9.1

# sentry
raven[flask]

# other requirements
blinker
PyYAML
requests
